+++ CHANGELOG.rst
@@ -1,5 +1,13 @@
-Will be updated by antsibull-changelog. Do not edit this manually!
+=============================================
+MacOS SoftwareUpdate Collection Release Notes
+=============================================
 
-See https://ansible.readthedocs.io/projects/antsibull-changelog/changelogs/ for information on how to use antsibull-changelog.
+.. contents:: Topics
 
-Check out ``changelogs/config.yaml`` for its configuration. You need to change at least the ``title`` field in there.
+v1.0.1
+======
+
+Release Summary
+---------------
+
+Project initiated